+++ wp-content/uploads/civicrm/ext/mjwshared/CRM/Core/Payment/MJWIPNTrait.php
@@ -9,6 +9,7 @@
  +--------------------------------------------------------------------+
  */
 
+use Civi\Api4\Contribution;
 use Civi\Payment\Exception\PaymentProcessorException;
 
 /**
@@ -125,7 +126,7 @@ trait CRM_Core_Payment_MJWIPNTrait {
     // Non-recurring contribution.
     else {
       try {
-        $this->is_email_receipt = \Civi\Api4\Contribution::get(FALSE)
+        $this->is_email_receipt = Contribution::get(FALSE)
           ->addSelect('contribution_page_id.is_email_receipt')
           ->addWhere('id', '=', $contributionID)
           ->execute()
@@ -293,9 +294,11 @@ trait CRM_Core_Payment_MJWIPNTrait {
    */
   private function updateRecurCompleted($params) {
     $this->checkRequiredParams('updateRecurCompleted', ['id'], $params);
-    $params['contribution_status_id'] = 'Completed';
 
-    civicrm_api3('ContributionRecur', 'create', $params);
+    \Civi\Api4\ContributionRecur::update(FALSE)
+      ->addWhere('id', '=', $params['id'])
+      ->addValue('contribution_status_id:name', 'Completed')
+      ->execute();
   }
 
   /**
@@ -324,13 +327,13 @@ trait CRM_Core_Payment_MJWIPNTrait {
   /**
    * Repeat a contribution (call the Contribution.repeattransaction API)
    *
-   * @param array $params
+   * @param array $repeatContributionParams
    *
-   * @return bool
+   * @return int
    * @throws \CiviCRM_API3_Exception
    * @throws \Civi\Payment\Exception\PaymentProcessorException
    */
-  private function repeatContribution($params) {
+  private function repeatContribution(array $repeatContributionParams): int {
     $this->checkRequiredParams(
       'repeatContribution',
       [
@@ -342,14 +345,14 @@ trait CRM_Core_Payment_MJWIPNTrait {
         'trxn_id',
         //'total_amount',
       ],
-      $params
+      $repeatContributionParams
     );
     // Optional Params: fee_amount
 
     // Creat contributionParams for Contribution.repeattransaction and set some values
-    $contributionParams = $params;
+    $contributionParams = $repeatContributionParams;
     // Status should be pending if we have a successful payment
-    switch ($params['contribution_status_id']) {
+    switch ($contributionParams['contribution_status_id']) {
       case CRM_Core_PseudoConstant::getKey('CRM_Contribute_BAO_Contribution', 'contribution_status_id', 'Completed'):
         // Create a contribution in Pending state using Contribution.repeattransaction and then complete using Payment.create
         $contributionParams['contribution_status_id'] = 'Pending';
@@ -359,12 +362,11 @@ trait CRM_Core_Payment_MJWIPNTrait {
       default:
         // Failed etc.
         // For any other status create it directly using Contribution.repeattransaction and don't create a payment
-        $contributionParams['contribution_status_id'] = $params['contribution_status_id'];
         $createPayment = FALSE;
     }
 
     // Create the new contribution
-    $contributionParams['trxn_id'] = $params['order_reference'];
+    $contributionParams['trxn_id'] = $repeatContributionParams['order_reference'];
     // We send a receipt when adding a payment, not now
     $contributionParams['is_email_receipt'] = FALSE;
     try {
@@ -378,28 +380,32 @@ trait CRM_Core_Payment_MJWIPNTrait {
     }
 
     // Get total amount from contribution returned by repeatTransaction (Which came from the recur template Contribution)
-    $params['total_amount'] = $params['total_amount'] ?? $contribution['total_amount'];
+    $repeatContributionParams['total_amount'] = $repeatContributionParams['total_amount'] ?? $contribution['total_amount'];
 
     if ($createPayment) {
-      $paymentParamsKeys = [
+      // Start by passing through all params (we may pass in customdata as well in API4 format)
+      $paymentParams = $repeatContributionParams;
+      // Now map contribution keys to payment keys
+      $paymentParamsMap = [
         'receive_date' => 'trxn_date',
         'order_reference' => 'order_reference',
         'trxn_id' => 'trxn_id',
         'total_amount' => 'total_amount',
         'fee_amount' => 'fee_amount',
       ];
-      foreach ($paymentParamsKeys as $contributionKey => $paymentKey) {
-        if (isset($params[$contributionKey])) {
-          $paymentParams[$paymentKey] = $params[$contributionKey];
+      foreach ($paymentParamsMap as $contributionKey => $paymentKey) {
+        if (isset($contributionParams[$contributionKey])) {
+          unset($paymentParams[$contributionKey]);
+          $paymentParams[$paymentKey] = $contributionParams[$contributionKey];
         }
       }
       $paymentParams['contribution_id'] = $contribution['id'];
-      $paymentParams['payment_processor_id'] = $this->_paymentProcessor->getID();
+      $paymentParams['payment_processor_id'] = $this->getPaymentProcessor()->getID();
       $paymentParams['is_send_contribution_notification'] = $this->getSendEmailReceipt($paymentParams['contribution_id']);
       $paymentParams['skipCleanMoney'] = TRUE;
       civicrm_api3('Mjwpayment', 'create_payment', $paymentParams);
     }
-    return TRUE;
+    return $contribution['id'];
   }
 
   /**
@@ -422,15 +428,18 @@ trait CRM_Core_Payment_MJWIPNTrait {
   /**
    * Complete a pending contribution and update associated entities (recur/membership)
    *
-   * @param array $params
+   * @param array $contributionParams
    *
    * @throws \CiviCRM_API3_Exception
    * @throws \Civi\Payment\Exception\PaymentProcessorException
    */
-  private function updateContributionCompleted($params) {
-    $this->checkRequiredParams('updateContributionCompleted', ['contribution_id', 'trxn_date', 'order_reference', 'trxn_id', 'total_amount'], $params);
+  private function updateContributionCompleted(array $contributionParams) {
+    $this->checkRequiredParams('updateContributionCompleted', ['contribution_id', 'trxn_date', 'order_reference', 'trxn_id', 'total_amount'], $contributionParams);
 
-    $paymentParamsKeys = [
+    // Start by passing through all params (we may pass in customdata as well in API4 format)
+    $paymentParams = $contributionParams;
+    // Now map contribution keys to payment keys
+    $paymentParamsMap = [
       'contribution_id' => 'contribution_id',
       'trxn_date' => 'trxn_date',
       'order_reference' => 'order_reference',
@@ -438,9 +447,10 @@ trait CRM_Core_Payment_MJWIPNTrait {
       'total_amount' => 'total_amount',
       'fee_amount' => 'fee_amount',
     ];
-    foreach ($paymentParamsKeys as $contributionKey => $paymentKey) {
-      if (isset($params[$contributionKey])) {
-        $paymentParams[$paymentKey] = $params[$contributionKey];
+    foreach ($paymentParamsMap as $contributionKey => $paymentKey) {
+      if (isset($contributionParams[$contributionKey])) {
+        unset($paymentParams[$contributionKey]);
+        $paymentParams[$paymentKey] = $contributionParams[$contributionKey];
       }
     }
 
@@ -451,17 +461,17 @@ trait CRM_Core_Payment_MJWIPNTrait {
     // Also set cancel_date to NULL because we are setting it for failed payments and UI will still display "greyed
     // out" if it is set.
     $failedContributionStatus = (int) CRM_Core_PseudoConstant::getKey('CRM_Contribute_BAO_Contribution', 'contribution_status_id', 'Failed');
-    if (isset($params['contribution_status_id']) && ((int) $params['contribution_status_id'] === $failedContributionStatus)) {
+    if (isset($contributionParams['contribution_status_id']) && ((int) $contributionParams['contribution_status_id'] === $failedContributionStatus)) {
       $sql = "UPDATE civicrm_contribution SET contribution_status_id=%1,cancel_date=NULL WHERE id=%2";
       $queryParams = [
         1 => [CRM_Core_PseudoConstant::getKey('CRM_Contribute_BAO_Contribution', 'contribution_status_id', 'Pending'), 'Positive'],
-        2 => [$params['contribution_id'], 'Positive'],
+        2 => [$contributionParams['contribution_id'], 'Positive'],
       ];
       CRM_Core_DAO::executeQuery($sql, $queryParams);
     }
     $paymentParams['is_send_contribution_notification'] = $this->getSendEmailReceipt($paymentParams['contribution_id']);
     $paymentParams['skipCleanMoney'] = TRUE;
-    $paymentParams['payment_processor_id'] = $this->_paymentProcessor->getID();
+    $paymentParams['payment_processor_id'] = $this->getPaymentProcessor()->getID();
     civicrm_api3('Mjwpayment', 'create_payment', $paymentParams);
   }
 
@@ -475,13 +485,13 @@ trait CRM_Core_Payment_MJWIPNTrait {
    */
   private function updateContributionFailed($params) {
     $this->checkRequiredParams('updateContributionFailed', ['contribution_id', 'order_reference'], $params);
-    civicrm_api3('Contribution', 'create', [
-      'contribution_status_id' => 'Failed',
-      'id' => $params['contribution_id'],
-      'trxn_id' => $params['order_reference'],
-      'cancel_date' => $params['cancel_date'] ?? NULL,
-      'cancel_reason' => $params['cancel_reason'] ?? NULL,
-    ]);
+    Contribution::update(FALSE)
+      ->addValue('contribution_status_id:name', 'Failed')
+      ->addValue('trxn_id', $params['order_reference'])
+      ->addValue('cancel_date', $params['cancel_date'] ?? '')
+      ->addValue('cancel_reason', $params['cancel_reason'] ?? '')
+      ->addWhere('id', '=', $params['contribution_id'])
+      ->execute();
     // No financial_trxn is created so we don't need to update that.
   }
 
@@ -567,7 +577,7 @@ trait CRM_Core_Payment_MJWIPNTrait {
    * @throws \Civi\Payment\Exception\PaymentProcessorException
    */
   protected function exception($message) {
-    $label = method_exists($this->_paymentProcessor, 'getPaymentProcessorLabel') ? $this->_paymentProcessor->getPaymentProcessorLabel() : __CLASS__;
+    $label = method_exists($this->getPaymentProcessor(), 'getPaymentProcessorLabel') ? $this->getPaymentProcessor()->getPaymentProcessorLabel() : __CLASS__;
     $errorMessage = $label . ' Exception: Event: ' . $this->eventType . ' Error: ' . $message;
     Civi::log()->error($errorMessage);
     if ($this->exitOnException) {